\name{LPKsample-package}
\alias{LPKsample-package}
\alias{LPKsample}
\docType{package}
\title{
\packageTitle{LPKsample}
}
\description{
    This package performs high dimensional K-sample comparison using graph-based LP nonparametric (GLP) method. 
}

\author{ Mukhopadhyay, S. and Wang, K.

Maintainer: \packageMaintainer{LPKsample}
}
\references{
Mukhopadhyay, S. and Wang, K. (2020), "A Nonparametric Approach to High-dimensional K-sample Comparison Problem", arXiv:1810.01724.

Mukhopadhyay, S. (2017+), "Unified Statistical Theory of Spectral Graph Analysis".

Mukhopadhyay, S. and Parzen, E. (2014), "LP Approach to Statistical Modeling", arXiv:1405.2601.
}

\keyword{ package }