\name{LPKsample-package}
\alias{LPKsample-package}
\alias{LPKsample}
\docType{package}
\title{
\packageTitle{LPKsample}
}
\description{
    This package performs high dimensional K-sample comparison using graph-based LP nonparametric (GLP) method. 
}

\author{ Mukhopadhyay, S. and Wang, K.

Maintainer: Kaijun Wang <kaijun.wang@temple.edu>
}
\references{
Mukhopadhyay, S. and Wang, K. (2017), "A Nonparametric Approach to High-dimensional K-sample Comparison Problem".

Mukhopadhyay, S. (2017+), "Unified Statistical Theory of Spectral Graph Analysis".

Mukhopadhyay, S. and Parzen, E. (2014), "LP Approach to Statistical Modeling", arXiv:1405.2601.
}

\keyword{ package }